start

+++ src/ruuvi.h
@@ -17,9 +17,9 @@ typedef struct
     uint8_t power;
     uint16_t sequence;
     uint8_t moveCount;
-    int ruuvi_format;
+    uint8_t ruuvi_format;
     int16_t rssi;
-    char name[30];
-    char address[20];
+        char name[30];
+        char address[20];
     uint32_t last_sequence;
 } ruuviAd_t;
